requirements.txt
streamlit
pandas
pandasai
openpyxl
PyYAML
